docs(README): add code example in `README.md`
follow the suggestion: <https://github.com/WSH032/pytauri/issues/2#issuecomment-2570432097>
