--- docs_src/tutorial/command.py
@@ -1,21 +1,0 @@
-import sys
-
-from pydantic import BaseModel
-from pytauri import Commands
-
-commands: Commands = Commands()
-
-
-class Person(BaseModel):
-    name: str
-
-
-class Greeting(BaseModel):
-    message: str
-
-
-@commands.command()
-async def greet(body: Person) -> Greeting:
-    return Greeting(
-        message=f"Hello, {body.name}! You've been greeted from Python {sys.version}!"
-    )
