+++ docs_src/tutorial/invoke_handler.py
@@ -1,4 +1,9 @@
+# --8<-- [start:command]
+
+import sys
+
 from anyio.from_thread import start_blocking_portal
+from pydantic import BaseModel
 from pytauri import (
     BuilderArgs,
     Commands,
@@ -7,6 +12,24 @@ from pytauri import (
 )
 
 commands: Commands = Commands()
+
+
+class Person(BaseModel):
+    name: str
+
+
+class Greeting(BaseModel):
+    message: str
+
+
+@commands.command()
+async def greet(body: Person) -> Greeting:
+    return Greeting(
+        message=f"Hello, {body.name}! You've been greeted from Python {sys.version}!"
+    )
+
+
+# --8<-- [end:command]
 
 
 def main() -> None:
